# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a complex Chain Cms.",
        "title": "Chain Cms Doc",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "paths": {
        "/attach/upload/blob": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "附件模块"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blob名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "blob扩展名",
                        "name": "extension",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "blob内容",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/attach.UploadBlobRes"
                        }
                    }
                }
            }
        },
        "/entity/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "pplication/json"
                ],
                "tags": [
                    "entity 模块"
                ],
                "summary": "新增 entity",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddEntityReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/entity/add/multi": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "pplication/json"
                ],
                "tags": [
                    "entity 模块"
                ],
                "summary": "批量新增 entity",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AddEntityMultiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/entity/factory/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entity 模块"
                ],
                "summary": "获取用户持有的 factory 总览",
                "parameters": [
                    {
                        "maxLength": 255,
                        "minLength": 1,
                        "type": "string",
                        "description": "拥有者",
                        "name": "possessor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.GetUserFactoryAllRes"
                        }
                    }
                }
            }
        },
        "/entity/factory/entity/all": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entity 模块"
                ],
                "summary": "获取用户持有的(某个 factory 下的)所有 entity",
                "parameters": [
                    {
                        "maxLength": 30,
                        "minLength": 1,
                        "type": "string",
                        "description": "链网络标识",
                        "name": "chainMagic",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 30,
                        "minLength": 1,
                        "type": "string",
                        "description": "链名",
                        "name": "chainName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string",
                        "description": "模板ID",
                        "name": "factoryId",
                        "in": "query"
                    },
                    {
                        "maxLength": 255,
                        "minLength": 1,
                        "type": "string",
                        "description": "拥有者",
                        "name": "possessor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 2,
                        "minimum": 1,
                        "type": "integer",
                        "description": "类型(1:普通,2:限量,默认:1)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.GetUserFactoryEntityAllRes"
                        }
                    }
                }
            }
        },
        "/entity/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "pplication/json"
                ],
                "tags": [
                    "entity 模块"
                ],
                "summary": "更新 entity",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateEntityReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "attach.UploadBlobRes": {
            "type": "object",
            "properties": {
                "url": {
                    "description": "文件路径",
                    "type": "string"
                }
            }
        },
        "entity.AddEntityMultiReq": {
            "type": "object",
            "required": [
                "chainMagic",
                "chainName",
                "entities",
                "factoryId",
                "possessor",
                "taxCollector",
                "type"
            ],
            "properties": {
                "chainMagic": {
                    "description": "链网络标识",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "chainName": {
                    "description": "链名",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "entities": {
                    "description": "非同质资产列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.EntityStruct"
                    }
                },
                "extension": {
                    "description": "blob扩展名",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "factoryId": {
                    "description": "模板ID",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "factoryName": {
                    "description": "模板名称",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "hash": {
                    "description": "blob哈希",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "possessor": {
                    "description": "拥有者",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "taxCollector": {
                    "description": "收税人",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "type": {
                    "description": "类型(1:普通,2:限量,默认:1)",
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 1
                }
            }
        },
        "entity.AddEntityReq": {
            "type": "object",
            "required": [
                "chainMagic",
                "chainName",
                "entityId",
                "factoryId",
                "possessor",
                "taxAssetPrealnum",
                "taxCollector",
                "type"
            ],
            "properties": {
                "chainMagic": {
                    "description": "链网络标识",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "chainName": {
                    "description": "链名",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "entityId": {
                    "description": "非同质资产ID",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "extension": {
                    "description": "blob扩展名",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "factoryId": {
                    "description": "模板ID",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "factoryName": {
                    "description": "模板名称",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "hash": {
                    "description": "blob哈希",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "possessor": {
                    "description": "拥有者",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "taxAssetPrealnum": {
                    "description": "缴纳数量",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "taxCollector": {
                    "description": "收税人",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "type": {
                    "description": "类型(1:普通,2:限量,默认:1)",
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 1
                }
            }
        },
        "entity.EntityStruct": {
            "type": "object",
            "required": [
                "entityId",
                "taxAssetPrealnum"
            ],
            "properties": {
                "entityId": {
                    "description": "非同质资产ID",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "taxAssetPrealnum": {
                    "description": "缴纳数量",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "taxCollector": {
                    "description": "收税人",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "entity.GetUserFactoryAllRes": {
            "type": "object",
            "properties": {
                "factories": {
                    "description": "factory 列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.UserFactoryInfo"
                    }
                }
            }
        },
        "entity.GetUserFactoryEntityAllRes": {
            "type": "object",
            "properties": {
                "entities": {
                    "description": "factory/entites 列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.UserFactoryEntityInfo"
                    }
                }
            }
        },
        "entity.SubEntity": {
            "type": "object",
            "properties": {
                "entityId": {
                    "description": "非同质资产ID",
                    "type": "string"
                },
                "taxAssetPrealnum": {
                    "description": "缴纳数量",
                    "type": "string"
                },
                "taxCollector": {
                    "description": "收税人",
                    "type": "string"
                }
            }
        },
        "entity.UpdateEntityReq": {
            "type": "object",
            "required": [
                "chainMagic",
                "chainName",
                "entityId",
                "factoryId"
            ],
            "properties": {
                "chainMagic": {
                    "description": "链网络标识",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "chainName": {
                    "description": "链名",
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "entityId": {
                    "description": "非同质资产ID",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "factoryId": {
                    "description": "模板ID",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "factoryName": {
                    "description": "模板名称",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "possessor": {
                    "description": "拥有者",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "entity.UserFactoryEntityInfo": {
            "type": "object",
            "properties": {
                "chainMagic": {
                    "description": "链网络标识",
                    "type": "string"
                },
                "chainName": {
                    "description": "链名",
                    "type": "string"
                },
                "entityId": {
                    "description": "非同质资产ID",
                    "type": "string"
                },
                "extension": {
                    "description": "blob扩展名",
                    "type": "string"
                },
                "factoryId": {
                    "description": "模板ID",
                    "type": "string"
                },
                "factoryName": {
                    "description": "模板名称",
                    "type": "string"
                },
                "hash": {
                    "description": "blob哈希",
                    "type": "string"
                },
                "possessor": {
                    "description": "拥有者",
                    "type": "string"
                },
                "subEntities": {
                    "description": "子集",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SubEntity"
                    }
                },
                "taxAssetPrealnum": {
                    "description": "缴纳数量",
                    "type": "string"
                },
                "taxCollector": {
                    "description": "收税人",
                    "type": "string"
                },
                "type": {
                    "description": "类型(1:普通,2:限量,默认:1)",
                    "type": "integer"
                }
            }
        },
        "entity.UserFactoryInfo": {
            "type": "object",
            "properties": {
                "chainMagic": {
                    "description": "链网络标识",
                    "type": "string"
                },
                "chainName": {
                    "description": "链名",
                    "type": "string"
                },
                "factoryId": {
                    "description": "模板ID",
                    "type": "string"
                },
                "factoryName": {
                    "description": "模板名称",
                    "type": "string"
                },
                "numberOfEntities": {
                    "description": "持有的 entity 数量",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add prefix \"Bearer \", such as \"Bearer token\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}